corrected PR comments
